DATAGRAPH-793 - Upgrade to querydsl 4.0.x.
We upgraded to the latest querydsl version that is officially supported.
The com.mysema.querydsl project moved to a different organization 
which required a package change.

Adapted dependency declarations as well as classnames.
I also needed to add the messing json-simple dependency.
Currently using a non official version of the neo4j-cypher-dsl
that I also adapted to the latest querydsl version:
https://github.com/neo4j-contrib/cypher-dsl/pull/50

With this in place Spring Data Neo4J can be built again :).